feat: return result when calling workflow explicitly
